ragas_test/requirements.txt
ragas>=0.2.14
langchain>=0.1.0
langchain-openai>=0.0.2
faiss-cpu>=1.7.4
openai>=1.2.0
datasets>=2.0.0
pandas>=1.3.0
numpy>=1.20.0
python-dotenv>=1.0.0
transformers>=4.33.0
tiktoken>=0.5.1
argparse>=1.4.0 